yz3301 Merge_Worksheet Submission
1. In your own words, what is version (or source) control?
Version control is a systematic category of software tools that facilitate a software team to collectively modify source code over time. The version control system keeps track of the modifications made over time. If there is any mistakes or unwanted changes, the master player could go back and make change while minimizing disruption for other players.

2. Explain the difference between Git and GitHub.
Git is a charge-free source that constitutes version-control systems with speed, efficiency, and a hinted democratic enlightenment agenda.
GitHub is a hosting service for Git repositories.

3. In a few sentences, explain what a "commit" is?
“Commit” in version control systems, refers to the latest changes made to source code repository. Each commit is restored in the system database permanently

4.What is a "branch" in Git terms?
“Branch” refers to a movable pointer of one or more of the “commits” (changes). The "subsidized branch" makes request to merge the changes to the “default branch”

5.What is a repository? Explain what happens when you "push" to and "pull" from a repository.
Since the purpose of “Git” is to store information and the modifications made to those information over time. The storage system in “Git” is called “repository.” In Git systems, developers need to pull stored information to their local repositories; or manually push commits to the master branch.

6.What is involved in the "clone" operation?
The “clone” operation makes copies of everything in existing Git repository to a local drive that is completely isolated from the original Git repository.

7.Find an example of Git being used for a peer production project. In several paragraphs answer the following: Who are they? What do they do? Who maintains the repository? What are the rules for contributing?

“Towards the Right” (afc163) is ranked as the top fifteen most active user on GitHub. This is a group is affiliated to the AliPay department of the Chinese Corporate Ali Baba (the CEO is Jack Ma).  Members of “Towards the Right” is dedicated to modifying and perfecting the user-computer interface of AliPay Programs. Modifications made to the AliPay Software are presented in this repository for mainly two reasons:
1. To maintain and marketize the brand of AliPay by showing how the Alipay team is dedicated to even delicate details of the software.
2. The files contained in the repository slightly and mildly criticize the rigid hierarchal structure and inflexible nature of the works at AliPay. As it is named after “Towards the Right,” this is metaphorically referring that the work philosophy in AliPay is similar to the CCP’s Leftist and rigid structures.
Even if it is not clearly stated, I assume that this group is organized by department authorities since the files are selectively uploaded to the master branch for public access. The rules, from my guess, may contain points like any contaminate the reputation for AliPay. Don’t go directly and furiously against the command of CCP.

8. Submit this worksheet (in .txt or .md plain text format) by filing a pull request to our Github repository.